refactor deletion code
