+++ packages/db/src/index.ts
@@ -73,7 +73,6 @@ export { translateApiGranuleToPostgresGranule } from './translate/granules';
 
 export {
   createGranuleWithExecutionHistory,
-  deleteGranuleWithExecutionHistory,
   upsertGranuleWithExecutionHistory,
 } from './lib/granule';
